DTO 클래스 변경 및 요청/응답 핸들러 수정
이 커밋에서는 DTO 클래스의 구조를 변경하고, 이에 따라 요청/응답 핸들러를 수정하였습니다.

주요 변경사항:
1. CalculationListRequestDto와 CalculationListResponseDto의 역할이 바뀌었습니다. CalculationListRequestDto는 이제 개별 계산 결과를 나타내는 DTO로 사용되며, CalculationListResponseDto는 이제 계산 결과 리스트를 나타내는 DTO로 사용됩니다.
2. ListCalculationRequest와 PostCalculationRequest 클래스에서 CalculationListRequestDto 대신 CalculationListResponseDto를 사용하도록 수정하였습니다.
3. CalculationListRequestDto와 CalculationListResponseDto의 필드와 메소드가 변경되었습니다. CalculationListRequestDto는 이제 개별 계산 결과를 나타내는 필드를 가지며, CalculationListResponseDto는 계산 결과 리스트를 나타내는 필드를 가집니다.
