+++ app/src/main/java/com/example/demo/presentation/ListCalculationRequest.java
@@ -2,7 +2,7 @@ package com.example.demo.presentation;
 
 import com.example.demo.application.Calculator;
 import com.example.demo.infrastructure.Calculation;
-import com.example.demo.presentation.dto.CalculationListRequestDto;
+import com.example.demo.presentation.dto.CalculationListResponseDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.util.List;
@@ -18,7 +18,7 @@ public class ListCalculationRequest implements RequestMethodHandler {
         List<Calculation> calculationList = calculator.getCalculationList();
 
         return objectMapper.writeValueAsString(
-                CalculationListRequestDto.of(calculationList)
+                CalculationListResponseDto.of(calculationList)
         );
     }
 }
